+++ Packages/com.adapty.unity-sdk/Runtime/Models/AdaptyUIUserAction.cs
@@ -9,9 +9,11 @@ namespace AdaptySDK {
     public partial class AdaptyUIUserAction {
         public AdaptyUIUserActionType Type;
         public string Value;
+        public AdaptyWebPresentation? OpenIn;
 
         public override string ToString() =>
             $"{nameof(Type)}: {Type}, " +
-            $"{nameof(Value)}: {Value}";
+            $"{nameof(Value)}: {Value}, " +
+            $"{nameof(OpenIn)}: {OpenIn}";
     }
 }